Refactor MessageDetailsViewModel
